+++ src/estimagic/optimization/optimize.py
@@ -24,6 +24,7 @@ from estimagic.optimization.tiktak import run_multistart_optimization
 from estimagic.optimization.tiktak import WEIGHT_FUNCTIONS
 from estimagic.parameters.conversion import aggregate_func_output_to_value
 from estimagic.parameters.conversion import get_converter
+from estimagic.parameters.nonlinear_constraints import process_nonlinear_constraints
 from estimagic.process_user_function import process_func_of_params
 
 
@@ -515,6 +516,22 @@ def _optimize(
             raise ValueError(msg.format(algo_info.name))
 
     # ==================================================================================
+    # Split constraints into nonlinear and reparametrization parts
+    # ==================================================================================
+    if isinstance(constraints, dict):
+        constraints = [constraints]
+
+    nonlinear_constraints = [c for c in constraints if c["type"] == "nonlinear"]
+
+    if nonlinear_constraints and "nonlinear_constraints" not in algo_kwargs:
+        raise ValueError(
+            f"Algorithm {algo_info.name} does not support nonlinear constraints."
+        )
+
+    # the following constraints will be handled via reparametrization
+    constraints = [c for c in constraints if c["type"] != "nonlinear"]
+
+    # ==================================================================================
     # prepare logging
     # ==================================================================================
     if logging:
@@ -622,6 +639,7 @@ def _optimize(
         soft_upper_bounds=soft_upper_bounds,
         add_soft_bounds=multistart,
     )
+
     # ==================================================================================
     # initialize the log database
     # ==================================================================================
@@ -662,6 +680,15 @@ def _optimize(
         direction=direction,
     )
 
+    # process nonlinear constraints:
+    internal_constraints = process_nonlinear_constraints(
+        nonlinear_constraints=nonlinear_constraints,
+        params=params,
+        converter=converter,
+        numdiff_options=numdiff_options,
+        skip_checks=skip_checks,
+    )
+
     x = internal_params.values
     # ==================================================================================
     # get the internal algorithm
@@ -672,6 +699,7 @@ def _optimize(
         valid_kwargs=algo_kwargs,
         lower_bounds=internal_params.lower_bounds,
         upper_bounds=internal_params.upper_bounds,
+        nonlinear_constraints=internal_constraints,
         algo_options=algo_options,
         logging=logging,
         db_kwargs=db_kwargs,